consultation-agent/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
openai-whisper>=20231117
pyannote.audio>=3.3.2
google-genai>=1.30.0
reportlab>=4.2.2
motor>=3.7.1
pymongo>=4.15.0
pydantic>=2.8.2
requests>=2.32.0
